Queue and app status optimisations

+++ AM Downloader/OptionsViewModel.cs
@@ -21,20 +21,41 @@ namespace AMDownloader
 
         void SaveSettings(object obj)
         {
+            if (Settings.Default.MaxConnectionsPerDownload < 1 || Settings.Default.MaxConnectionsPerDownload > 5)
+            {
+                Settings.Default.MaxConnectionsPerDownload = 5;
+            }
+            if (Settings.Default.MaxParallelDownloads < 1 || Settings.Default.MaxParallelDownloads > 10)
+            {
+                Settings.Default.MaxParallelDownloads = 10;
+            }
             Settings.Default.Save();
         }
 
         void ResetSettings(object obj)
         {
             Settings.Default.Reset();
-            try
+            if (Directory.Exists(Common.ApplicationPaths.DownloadsHistory))
             {
-                Directory.Delete(Common.ApplicationPaths.DownloadsHistory,true);
-                Directory.Delete(Common.ApplicationPaths.SavedLocationsHistory,true);
+                try
+                {
+                    Directory.Delete(Common.ApplicationPaths.DownloadsHistory, true);
+                }
+                catch (Exception ex)
+                {
+                    Debug.WriteLine(ex.Message);
+                }
             }
-            catch
+            if (Directory.Exists(Common.ApplicationPaths.SavedLocationsHistory))
             {
-
+                try
+                {
+                    Directory.Delete(Common.ApplicationPaths.SavedLocationsHistory, true);
+                }
+                catch (Exception ex)
+                {
+                    Debug.WriteLine(ex.Message);
+                }
             }
         }
     }
